+++ src/main/java/com/kira/farm_fresh_store/request/order/OrderDetailDto.java
@@ -1,9 +1,5 @@
 package com.kira.farm_fresh_store.request.order;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.kira.farm_fresh_store.entity.order.Order;
-import com.kira.farm_fresh_store.entity.product.Product;
-import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,6 +12,6 @@ import lombok.Setter;
 public class OrderDetailDto {
     private String productId;
     private Integer quantity;
-    private Long price;
-    private Long totalPrice;
+    private Integer price;
+    private Integer totalPrice;
 }
